# contracts/openapi/backend.openapi.json
{
  "components": {
    "schemas": {
...
      "AuthorRunArtifactSummary": {
        "properties": {
          "artifact_key": {
            "default": "",
            "title": "Artifact Key",
            "type": "string"
          },
          "artifact_type": {
            "title": "Artifact Type",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "artifact_type",
          "updated_at"
        ],
        "title": "AuthorRunArtifactSummary",
        "type": "object"
      },
...
      "AuthorRunGetResponse": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/AuthorRunArtifactSummary"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Node"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "raw_brief": {
            "title": "Raw Brief",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "review_ready",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "story_id",
          "status",
          "raw_brief",
          "created_at",
          "updated_at"
        ],
        "title": "AuthorRunGetResponse",
        "type": "object"
      },
      "AuthorStoryGetResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "draft_pack": {
            "additionalProperties": true,
            "title": "Draft Pack",
            "type": "object"
          },
          "latest_published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published At"
          },
          "latest_published_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Version"
          },
          "latest_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthorRunGetResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "title",
          "created_at"
        ],
        "title": "AuthorStoryGetResponse",
        "type": "object"
      },
      "AuthorStoryListItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "latest_published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published At"
          },
          "latest_published_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Version"
          },
          "latest_run_current_node": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Current Node"
          },
          "latest_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Id"
          },
          "latest_run_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Status"
          },
          "latest_run_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Run Updated At"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "title",
          "created_at"
        ],
        "title": "AuthorStoryListItem",
        "type": "object"
      },
      "AuthorStoryListResponse": {
        "properties": {
          "stories": {
            "items": {
              "$ref": "#/components/schemas/AuthorStoryListItem"
            },
            "title": "Stories",
            "type": "array"
          }
        },
        "title": "AuthorStoryListResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StoryCreateRequest": {
        "properties": {
          "pack_json": {
            "additionalProperties": true,
            "title": "Pack Json",
            "type": "object"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "StoryCreateRequest",
        "type": "object"
      },
      "StoryCreateResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "status": {
            "const": "draft",
            "title": "Status",
            "type": "string"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "status",
          "created_at"
        ],
        "title": "StoryCreateResponse",
        "type": "object"
      },
      "StoryDraftGetResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "draft_pack": {
            "additionalProperties": true,
            "title": "Draft Pack",
            "type": "object"
          },
          "latest_published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published At"
          },
          "latest_published_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Version"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "title",
          "created_at",
          "draft_pack"
        ],
        "title": "StoryDraftGetResponse",
        "type": "object"
      },
      "StoryDraftPatchChange": {
        "properties": {
          "field": {
            "enum": [
              "title",
              "description",
              "style_guard",
              "input_hint",
              "scene_seed",
              "red_line",
              "intro_text",
              "goal_hint",
              "starter_prompt_1",
              "starter_prompt_2",
              "starter_prompt_3"
            ],
            "title": "Field",
            "type": "string"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_type": {
            "enum": [
              "story",
              "beat",
              "scene",
              "npc",
              "opening_guidance"
            ],
            "title": "Target Type",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "target_type",
          "field",
          "value"
        ],
        "title": "StoryDraftPatchChange",
        "type": "object"
      },
      "StoryDraftPatchRequest": {
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/StoryDraftPatchChange"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Changes",
            "type": "array"
          }
        },
        "required": [
          "changes"
        ],
        "title": "StoryDraftPatchRequest",
        "type": "object"
      },
      "StoryGetResponse": {
        "properties": {
          "pack": {
            "additionalProperties": true,
            "title": "Pack",
            "type": "object"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "story_id",
          "version",
          "pack"
        ],
        "title": "StoryGetResponse",
        "type": "object"
      },
      "StoryListItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "has_draft": {
            "title": "Has Draft",
            "type": "boolean"
          },
          "latest_published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published At"
          },
          "latest_published_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Version"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "story_id",
          "title",
          "created_at",
          "has_draft"
        ],
        "title": "StoryListItem",
        "type": "object"
      },
      "StoryListResponse": {
        "properties": {
          "stories": {
            "items": {
              "$ref": "#/components/schemas/StoryListItem"
            },
            "title": "Stories",
            "type": "array"
          }
        },
        "title": "StoryListResponse",
        "type": "object"
      },
      "StoryPublishResponse": {
        "properties": {
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "story_id": {
            "title": "Story Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "story_id",
          "version",
          "published_at"
        ],
        "title": "StoryPublishResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "RPG Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/author/stories": {
      "get": {
        "operationId": "list_author_stories_endpoint_author_stories_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorStoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Author Stories Endpoint",
        "tags": [
          "author"
        ]
      }
    },
    "/author/stories/{story_id}": {
      "get": {
        "operationId": "get_author_story_endpoint_author_stories__story_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorStoryGetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Author Story Endpoint",
        "tags": [
          "author"
        ]
      }
    },
...
    "/stories": {
      "get": {
        "operationId": "list_stories_endpoint_stories_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Stories Endpoint",
        "tags": [
          "stories"
        ]
      },
      "post": {
        "operationId": "create_story_endpoint_stories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Story Endpoint",
        "tags": [
          "stories"
        ]
      }
    },
    "/stories/{story_id}": {
      "get": {
        "operationId": "get_story_endpoint_stories__story_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryGetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Story Endpoint",
        "tags": [
          "stories"
        ]
      }
    },
    "/stories/{story_id}/draft": {
      "get": {
        "operationId": "get_story_draft_endpoint_stories__story_id__draft_get",
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryDraftGetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Story Draft Endpoint",
        "tags": [
          "stories"
        ]
      },
      "patch": {
        "operationId": "patch_story_draft_endpoint_stories__story_id__draft_patch",
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StoryDraftPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryDraftGetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Story Draft Endpoint",
        "tags": [
          "stories"
        ]
      }
    },
    "/stories/{story_id}/publish": {
      "post": {
        "operationId": "publish_story_endpoint_stories__story_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "story_id",
            "required": true,
            "schema": {
              "title": "Story Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoryPublishResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Story Endpoint",
        "tags": [
          "stories"
        ]
      }
    }
  }
}